fixed notes
